VisualWallet-backend/mysite/requirements.txt
opencv-python
python-bitcoinlib
qrcode
pillow
numpy
scipy